# api-document/resources/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/is/v1/comment-service/add": {
            "post": {
                "description": "insert into table comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COMMENT SERVICE"
                ],
                "summary": "AddComment",
                "operationId": "AddComment",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-handlers_AddCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-handlers_AddCommentResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/comment-service/get-comment": {
            "post": {
                "description": "select * from table comment where productID = :1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COMMENT SERVICE"
                ],
                "summary": "GetCommentsByProductID",
                "operationId": "GetCommentsByProductID",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-handlers_GetCommentsByProductIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-handlers_GetCommentsByProductIDResponse"
                        }
                    }
                }
            }
        },
...
        "/is/v1/product-service/get-product-detail": {
            "post": {
                "description": "Get the detail of a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "GetProductDetail",
                "operationId": "GetProductDetail",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_product_GetProductDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_product_GetProductDetailResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/product-service/get-products": {
            "post": {
                "description": "list all the products in the inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "GetAllProduct",
                "operationId": "GetAllProduct",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_product_GetAllProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_product_GetAllProductResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_kingstonduy_api-document_domain_product.GetAllProductRequest": {
            "type": "object"
        },
        "github_com_kingstonduy_api-document_domain_product.GetAllProductResponse": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.Product"
                    }
                }
            }
        },
        "github_com_kingstonduy_api-document_domain_product.GetProductDetailRequest": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                }
            }
        },
        "github_com_kingstonduy_api-document_domain_product.GetProductDetailResponse": {
            "type": "object",
            "properties": {
                "productCatergory": {
                    "type": "string"
                },
                "productDescription": {
                    "type": "string"
                },
                "productId": {
                    "type": "string"
                },
                "productImage": {
                    "type": "string"
                },
                "productName": {
                    "type": "string"
                },
                "productPrice": {
                    "type": "number"
                }
            }
        },
        "github_com_kingstonduy_api-document_domain_product.Product": {
            "type": "object",
            "properties": {
                "averageRating": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "handlers.AddCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "productId": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "handlers.AddCommentResponse": {
            "type": "object"
        },
...
        "handlers.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                },
                "userImage": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
...
        "handlers.GetCommentsByProductIDRequest": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "string"
                }
            }
        },
        "handlers.GetCommentsByProductIDResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Comment"
                    }
                }
            }
        },
...
        "transport.Request-github_com_kingstonduy_api-document_domain_product_GetAllProductRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetAllProductRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_product_GetProductDetailRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetProductDetailRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Request-handlers_AddCommentRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.AddCommentRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Request-handlers_GetCommentsByProductIDRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.GetCommentsByProductIDRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Response-github_com_kingstonduy_api-document_domain_product_GetAllProductResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetAllProductResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_product_GetProductDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetProductDetailResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Response-handlers_AddCommentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.AddCommentResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Response-handlers_GetCommentsByProductIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.GetCommentsByProductIDResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "transport.Trace": {
            "type": "object",
            "required": [
                "cid",
                "cts"
            ],
            "properties": {
                "cid": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "cts": {
                    "type": "integer"
                },
                "dur": {
                    "type": "integer"
                },
                "frm": {
                    "type": "string"
                },
                "replyTo": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                },
                "sts": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "transactionTimeout": {
                    "type": "integer",
                    "maximum": 120000
                },
                "userName": {
                    "type": "string"
                }
            }
        }
    }
}